Use correct data type for rank in ModuleShort

+++ nerdd_backend/models/module.py
@@ -26,7 +26,7 @@ class ModulePublic(Module):
 
 class ModuleShort(BaseModel):
     id: str
-    rank: Optional[int] = None
+    rank: Optional[float] = None
     name: Optional[str] = None
     version: Optional[str] = None
     visible_name: Optional[str] = None
